#115 move signature checking to sepparate file
